+++ internal/wx/include/wx/msw/listbox.h
@@ -4,7 +4,7 @@
 // Author:      Julian Smart
 // Modified by:
 // Created:     01/02/97
-// RCS-ID:      $Id: listbox.h,v 1.47 2006/03/09 12:46:24 VZ Exp $
+// RCS-ID:      $Id: listbox.h,v 1.48 2006/03/14 19:35:26 ABX Exp $
 // Copyright:   (c) Julian Smart
 // Licence:     wxWindows licence
 /////////////////////////////////////////////////////////////////////////////
@@ -81,7 +81,7 @@ public:
     virtual void Clear();
     virtual void Delete(int n);
 
-    virtual int GetCount() const;
+    virtual size_t GetCount() const;
     virtual wxString GetString(int n) const;
     virtual void SetString(int n, const wxString& s);
     virtual int FindString(const wxString& s, bool bCase = false) const;
@@ -147,7 +147,7 @@ protected:
     // free memory (common part of Clear() and dtor)
     void Free();
 
-    int m_noItems;
+    size_t m_noItems;
     int m_selected;
 
     virtual wxSize DoGetBestSize() const;
